get all forms in the same file
