+++ libs/movie/src/lib/movie/form/budget/budget.form.ts
@@ -1,8 +1,5 @@
-import { FormEntity, FormValue, FormList } from '@blockframes/utils/form';
-import { FormControl, Validators } from '@angular/forms';
-import { BoxOffice } from '../../+state/movie.firestore';
+
 import { NumberRange } from '@blockframes/utils/common-interfaces/range';
-import { createMovie, Movie } from '@blockframes/movie/+state';
 
 export const BUDGET_LIST: NumberRange[] = [
   { from: 0, to: 1000000, label: 'Less than $1 million' },
@@ -13,47 +10,3 @@ export const BUDGET_LIST: NumberRange[] = [
   { from: 10000000, to: 20000000, label: '$10 - 20 millions' },
   { from: 20000000, to: 999999999, label: 'More than $20 millions' },
 ];
-
-function createBudgetFormControl(entity?: Partial<Movie>) {
-  const { boxOffice } = createMovie(entity);
-  return {
-    boxOffice: FormList.factory(boxOffice, el => new BoxOfficeForm(el))
-  }
-}
-
-export type BudgetFormControl = ReturnType<typeof createBudgetFormControl>;
-
-export class MovieBudgetForm extends FormEntity<BudgetFormControl> {
-  constructor(budget?: Movie) {
-    super(createBudgetFormControl(budget));
-  }
-}
-
-// Box Office
-
-function createBoxOfficeFormControl(boxOffice?: Partial<BoxOffice>) {
-  const { unit, territory, value } = createBoxOffice(boxOffice);
-  return {
-    unit: new FormValue(unit),
-    territory: new FormControl(territory),
-    value: new FormControl(value, Validators.min(0))
-  }
-}
-
-export type BoxOfficeFormControl = ReturnType<typeof createBoxOfficeFormControl>;
-
-export class BoxOfficeForm extends FormEntity<BoxOfficeFormControl> {
-  constructor(boxOffice?: Partial<BoxOffice>) {
-    super(createBoxOfficeFormControl(boxOffice))
-  }
-}
-
-export function createBoxOffice(params: Partial<BoxOffice> = {}): BoxOffice {
-  return {
-    unit: 'boxoffice_dollar',
-    value: 0,
-    territory: null,
-    ...params,
-  }
-}
-
