+++ libs/movie/src/lib/movie/form/movie.form.ts
@@ -1,24 +1,46 @@
+import {
+  MovieStakeholders,
+  MovieLegalDocuments,
+  BoxOffice,
+  Prize,
+  MoviePromotionalElements,
+  PromotionalHostedMedia,
+  PromotionalExternalMedia,
+  MovieReview,
+  MovieOriginalRelease,
+  MovieRating,
+  MovieLanguageSpecification,
+} from '../+state/movie.firestore';
+import {
+  Movie,
+  Credit,
+  createMovie,
+  createMovieLegalDocuments,
+  createTitle,
+  createStoreConfig,
+  createMovieStakeholders,
+  createMoviePromotional,
+  createPromotionalExternalMedia,
+  createPromotionalHostedMedia,
+  createMovieLanguageSpecification,
+} from '../+state/movie.model';
+
 import { FormArray, FormControl, Validators } from '@angular/forms';
-import { MoviePromotionalElementsForm, MoviePromotionalHostedMediaForm } from './promotional-elements/promotional-elements.form';
-import { Movie, createMovie, createMovieLegalDocuments } from '../+state';
 import { LegalDocument } from '@blockframes/contract/contract/+state/contract.firestore';
 import { FormStaticValue, FormStaticArray } from '@blockframes/utils/form/forms/static-value.form';
 import { createLegalDocument } from '@blockframes/contract/contract/+state/contract.model';
-import { MovieLegalDocuments } from '../+state/movie.firestore';
 import { FormEntity, EntityControl } from '@blockframes/utils/form/forms/entity.form';
 import { FormList } from '@blockframes/utils/form/forms/list.form';
-import { HostedMediaForm } from '@blockframes/media/form/media.form';
-import { MovieTechnicalInfoForm } from './techincal-info.form';
-import { TitleForm, DirectorForm, StoreConfigForm, StakeholderMapForm } from './main/main.form';
-import { yearValidators } from '@blockframes/utils/form/validators/validators';
+import { HostedMediaForm, ExternalMediaForm } from '@blockframes/media/form/media.form';
+import { yearValidators, urlValidators } from '@blockframes/utils/form/validators/validators';
 import { PriceForm } from '@blockframes/contract/version/form/price/price.form';
-import { BoxOfficeForm } from './budget/budget.form';
-import { MoviePrizeForm } from './festival-prizes/festival-prizes.form';
-import { MovieReviewForm } from './review/review.form';
-import { CreditForm } from './sales-cast/sales-cast.form';
-import { MovieRatingForm, OriginalReleaseForm } from './sales-info/sales-info.form';
-import { MovieVersionInfoForm } from './version-info/version-info.form';
-import { createLanguageControl } from '@blockframes/movie/form/version-info/version-info.form';
+import { FormValue } from '@blockframes/utils/form';
+import { createCredit, Stakeholder, createStakeholder } from '@blockframes/utils/common-interfaces/identity';
+import { createMovieAppAccess } from '@blockframes/utils/apps';
+import { MediaFormList } from '@blockframes/media/form/media-list.form';
+import { toDate } from '@blockframes/utils/helpers';
+import { LanguagesSlug } from '@blockframes/utils/static-model';
+
 
 // LEGAL DOCUMENTS
 
@@ -106,7 +128,6 @@ function createMovieControls(movie: Partial<Movie>) {
 
 export type MovieControl = ReturnType<typeof createMovieControls>;
 
-// @Injectable()
 export class MovieForm extends FormEntity<MovieControl, Movie> {
   constructor(movie?: Partial<Movie>) {
     super(createMovieControls(movie));
@@ -188,6 +209,88 @@ export class MovieForm extends FormEntity<MovieControl, Movie> {
     return this.get('synopsis');
   }
 
+  // DIRECTORS
+  public addDirector(credit?: Partial<Credit>): void {
+    const entity = createCredit(credit);
+    const creditControl = new DirectorForm(entity);
+    this.directors.push(creditControl);
+  }
+
+  public removeDirector(i: number): void {
+    this.directors.removeAt(i);
+  }
+
+  // PRIZES
+  public addPrize(): void {
+    const credit = new MoviePrizeForm();
+    this.prizes.push(credit);
+  }
+
+  public removePrize(i: number): void {
+    this.prizes.removeAt(i);
+  }
+
+  // CREDITS
+  public addCredit(credit?: Partial<Credit>, type: 'cast' | 'crew' | 'producer' = 'cast'): void {
+    switch (type) {
+      case 'producer':
+        const producer = createCredit(credit);
+        const producerControl = new CreditForm(producer);
+        this.producers.push(producerControl);
+        break;
+      case 'crew':
+        const crew = createCredit(credit);
+        const crewControl = new CreditForm(crew);
+        this.crew.push(crewControl);
+        break;
+      case 'cast':
+      default:
+        const cast = createCredit(credit);
+        const castControl = new CreditForm(cast);
+        this.cast.push(castControl);
+        break;
+    }
+  }
+
+  public removeCredit(i: number, type: 'cast' | 'crew' | 'producer' = 'cast'): void {
+    switch (type) {
+      case 'producer':
+        this.producers.removeAt(i);
+        break;
+      case 'crew':
+        this.crew.removeAt(i);
+        break;
+      case 'cast':
+      default:
+        this.cast.removeAt(i);
+        break;
+    }
+  }
+
+  // RATING
+  public addRating(): void {
+    const rating = new MovieRatingForm();
+    this.rating.push(rating);
+  }
+
+  public removeRating(i: number): void {
+    this.rating.removeAt(i);
+  }
+
+  // ORIGINAL RELEASE
+  public getOriginalRelease(i: number) {
+    return this.originalRelease.controls[i];
+  }
+
+  public addOriginalRelease(): void {
+    const orignialRelease = new OriginalReleaseForm();
+    this.originalRelease.push(orignialRelease);
+  }
+
+  public removeOriginalRelease(i: number): void {
+    this.originalRelease.removeAt(i);
+  }
+
   reset(value?: EntityControl<Movie>, options?: {
     onlySelf?: boolean;
     emitEvent?: boolean;
@@ -206,6 +309,430 @@ export class MovieForm extends FormEntity<MovieControl, Movie> {
       if (abstractControl instanceof FormArray) {
         abstractControl.clear();
       }
+    });
+  }
+}
+
+// ------------------------------
+//         BOX OFFICE
+// ------------------------------
+
+function createBoxOfficeFormControl(boxOffice?: Partial<BoxOffice>) {
+  const { unit, territory, value } = createBoxOffice(boxOffice);
+  return {
+    unit: new FormValue(unit),
+    territory: new FormControl(territory),
+    value: new FormControl(value, Validators.min(0))
+  }
+}
+
+export type BoxOfficeFormControl = ReturnType<typeof createBoxOfficeFormControl>;
+
+export class BoxOfficeForm extends FormEntity<BoxOfficeFormControl> {
+  constructor(boxOffice?: Partial<BoxOffice>) {
+    super(createBoxOfficeFormControl(boxOffice))
+  }
+}
+
+export function createBoxOffice(params: Partial<BoxOffice> = {}): BoxOffice {
+  return {
+    unit: 'boxoffice_dollar',
+    value: 0,
+    territory: null,
+    ...params,
+  }
+}
+
+// ------------------------------
+//         PRIZES
+// ------------------------------
+
+function createPrizeFormControl(entity?: Partial<Prize>) {
+  const { name, year, prize, logo, premiere } = createPrize(entity);
+  return {
+    name: new FormControl(name),
+    year: new FormControl(year, [yearValidators]),
+    prize: new FormControl(prize),
+    premiere: new FormControl(premiere),
+  }
+}
+
+type PrizeFormControl = ReturnType<typeof createPrizeFormControl>;
+
+export class MoviePrizeForm extends FormEntity<PrizeFormControl> {
+  constructor(prize?: Partial<Prize>) {
+    super(createPrizeFormControl(prize));
+  }
+}
+
+export function createPrize(prize: Partial<Prize> = {}): Prize {
+  return {
+    name: '',
+    year: null,
+    prize: '',
+    ...prize
+  };
+}
+
+// ------------------------------
+//          CREDIT
+// ------------------------------
+
+function createCreditFormControl(credit?: Partial<Credit>) {
+  const { firstName, lastName, role, status, description, filmography } = createCredit(credit);
+  return {
+    firstName: new FormControl(firstName),
+    lastName: new FormControl(lastName),
+    role: new FormControl(role),
+    filmography: new FormControl(filmography),
+    description: new FormControl(description),
+    status: new FormControl(status),
+  }
+}
+
+export type CreditFormControl = ReturnType<typeof createCreditFormControl>;
+
+export class CreditForm extends FormEntity<CreditFormControl> {
+  constructor(credit?: Partial<Credit>) {
+    super(createCreditFormControl(credit));
+  }
+}
+
+// ------------------------------
+//         DIRECTORS
+// ------------------------------
+
+export class DirectorForm extends FormEntity<DirectorFormControl> {
+  constructor(director?: Partial<Credit>) {
+    super(createDirectorFormControl(director))
+  }
+}
+
+function createDirectorFormControl(director?: Partial<Credit>) {
+  const { firstName, lastName, filmography, status, description } = createCredit(director);
+  return {
+    firstName: new FormControl(firstName, Validators.required),
+    lastName: new FormControl(lastName, Validators.required),
+    filmography: new FormControl(filmography),
+    description: new FormControl(description),
+    status: new FormControl(status),
+  }
+}
+
+type DirectorFormControl = ReturnType<typeof createDirectorFormControl>;
+
+// ------------------------------
+//       STAKEHOLDERS
+// ------------------------------
+
+export class StakeholderForm extends FormEntity<StakeholderControl, Stakeholder> {
+  constructor(stakeholder?: Partial<Stakeholder>) {
+    super(createStakeholderControl(stakeholder))
+  }
+}
+
+function createStakeholderControl(stakeholder?: Partial<Stakeholder>) {
+  const { displayName, countries } = createStakeholder(stakeholder);
+  return {
+    displayName: new FormControl(displayName),
+    countries: FormList.factory(countries, e => new FormStaticValue(e, 'TERRITORIES'))
+  }
+}
+
+type StakeholderControl = ReturnType<typeof createStakeholderControl>;
+
+
+// ------------------------------
+//       STAKEHOLDERS MAP
+// ------------------------------
+
+export class StakeholderMapForm extends FormEntity<StakeholderMapControl> {
+  constructor(stakeholders?: Partial<MovieStakeholders>) {
+    super(createStakeholderMapControl(stakeholders));
+  }
+}
+
+function createStakeholderMapControl(stakeholders?: Partial<MovieStakeholders>): StakeholderMapControl {
+  const entity = createMovieStakeholders(stakeholders);
+  const control = {};
+  for (const key in entity) {
+    control[key] = FormList.factory(entity[key], e => new StakeholderForm(e))
+  };
+  return control as StakeholderMapControl;
+}
+
+type StakeholderMapControl = {
+  [key in keyof MovieStakeholders]: FormList<Stakeholder, StakeholderForm>
+};
+
+// ------------------------------
+//         TITLE
+// ------------------------------
+
+export class TitleForm extends FormEntity<TitleFormControl> {
+  constructor(title?: Movie['title']) {
+    super(createTitleFormControl(title));
+  }
+}
+
+function createTitleFormControl(title?: Partial<Movie['title']>) {
+  const { original, international } = createTitle(title);
+  return {
+    original: new FormControl(original),
+    international: new FormControl(international, Validators.required),
+  }
+}
+
+type TitleFormControl = ReturnType<typeof createTitleFormControl>;
+
+// ------------------------------
+//       STORE CONFIG
+// ------------------------------
+
+export class StoreConfigForm extends FormEntity<StoreConfigControl> {
+  constructor(storeConfig?: Partial<Movie['storeConfig']>) {
+    super(createStoreConfigFormControl(storeConfig));
+  }
+}
+
+function createStoreConfigFormControl(storeConfig?: Partial<Movie['storeConfig']>) {
+  const { appAccess, status, storeType } = createStoreConfig(storeConfig);
+  return {
+    appAccess: new AppAccessForm(appAccess),
+    status: new FormControl(status),
+    storeType: new FormControl(storeType),
+  }
+}
+
+type StoreConfigControl = ReturnType<typeof createStoreConfigFormControl>;
+
+// ------------------------------
+//         APP ACCESS
+// ------------------------------
+
+export class AppAccessForm extends FormEntity<AppAccessControl> {
+  constructor(appAccess?: Partial<Movie['storeConfig']['appAccess']>) {
+    super(createAppAccessFormControl(appAccess));
+  }
+}
+
+function createAppAccessFormControl(appAccess?: Partial<Movie['storeConfig']['appAccess']>) {
+  const { catalog, festival } = createMovieAppAccess(appAccess);
+  return {
+    catalog: new FormControl(catalog),
+    festival: new FormControl(festival)
+  }
+}
+
+type AppAccessControl = ReturnType<typeof createAppAccessFormControl>;
+
+// ------------------------------
+//   Promotional External Media
+// ------------------------------
+
+function createPromotionalExternalMediaControl(promotionalExternalMedia?: Partial<PromotionalExternalMedia>) {
+  const { label, media } = createPromotionalExternalMedia(promotionalExternalMedia);
+  return {
+    label: new FormControl(label),
+    media: new ExternalMediaForm(media),
+  }
+}
+
+export type PromotionalExternalMediaControl = ReturnType<typeof createPromotionalExternalMediaControl>;
+
+export class MoviePromotionalExternalMediaForm extends FormEntity<PromotionalExternalMediaControl> {
+  constructor(promotionalExternalMedia?: Partial<PromotionalExternalMedia>) {
+    super(createPromotionalExternalMediaControl(promotionalExternalMedia));
+  }
+}
+
+// ------------------------------
+//   Promotional Hosted Media
+// ------------------------------
+
+function createPromotionalHostedMediaControl(promotionalHostedMedia?: Partial<PromotionalHostedMedia>) {
+  const { label, media } = createPromotionalHostedMedia(promotionalHostedMedia);
+  return {
+    label: new FormControl(label),
+    media: new HostedMediaForm(media),
+  }
+}
+export type PromotionalHostedMediaControl = ReturnType<typeof createPromotionalHostedMediaControl>;
+
+export class MoviePromotionalHostedMediaForm extends FormEntity<PromotionalHostedMediaControl> {
+  constructor(promotionalHostedMedia?: Partial<PromotionalHostedMedia>) {
+    super(createPromotionalHostedMediaControl(promotionalHostedMedia));
+  }
+}
+
+// ------------------------------
+//   Every Promotional Elements
+// ------------------------------
+
+function createMoviePromotionalElementsControls(promotionalElements?: Partial<MoviePromotionalElements>) {
+  const entity = createMoviePromotional(promotionalElements);
+
+  const stillPhotoControls: Record<string, MoviePromotionalHostedMediaForm> = {};
+  for (const key in entity.still_photo) {
+    stillPhotoControls[key] = new MoviePromotionalHostedMediaForm(entity.still_photo[key]);
+  }
+
+  return {
+    // Images
+    still_photo: new MediaFormList<Record<string, MoviePromotionalHostedMediaForm>>(stillPhotoControls),
+
+    // Hosted Media
+    presentation_deck: new MoviePromotionalHostedMediaForm(entity.presentation_deck),
+    scenario: new MoviePromotionalHostedMediaForm(entity.scenario),
+
+    // External Media
+    promo_reel_link: new MoviePromotionalExternalMediaForm(entity.promo_reel_link),
+    screener_link: new MoviePromotionalExternalMediaForm(entity.screener_link),
+    trailer_link: new MoviePromotionalExternalMediaForm(entity.trailer_link),
+    teaser_link: new MoviePromotionalExternalMediaForm(entity.teaser_link),
+  }
+}
+
+export type MoviePromotionalElementsControl = ReturnType<typeof createMoviePromotionalElementsControls>
+
+export class MoviePromotionalElementsForm extends FormEntity<MoviePromotionalElementsControl>{
+  constructor(promotionalElements?: MoviePromotionalElements) {
+    super(createMoviePromotionalElementsControls(promotionalElements));
+  }
+}
+
+// ------------------------------
+//           REVIEWS
+// ------------------------------
+
+function createMovieReviewControl(review: Partial<MovieReview> = {}) {
+  const { criticName, journalName, criticQuote, revueLink } = createMovieReview(review);
+  return {
+    criticName: new FormControl(criticName),
+    journalName: new FormControl(journalName),
+    criticQuote: new FormControl(criticQuote),
+    revueLink: new FormControl(revueLink, urlValidators),
+  }
+}
+
+export type MovieReviewControls = ReturnType<typeof createMovieReviewControl>;
+
+export class MovieReviewForm extends FormEntity<MovieReviewControls, MovieReview> {
+  constructor(review?: Partial<MovieReview>) {
+    super(createMovieReviewControl(review));
+  }
+}
+
+export function createMovieReview(params: Partial<MovieReview> = {}): MovieReview {
+  return {
+    criticName: '',
+    journalName: '',
+    criticQuote: '',
+    revueLink: '',
+    ...params,
+  }
+}
+
+// ------------------------------
+//           RATING
+// ------------------------------
+
+function createRatingFormControl(entity?: Partial<MovieRating>) {
+  const { country, reason, system, value } = createMovieRating(entity);
+  return {
+    country: new FormStaticValue(country, 'TERRITORIES'),
+    reason: new FormControl(reason),
+    system: new FormControl(system),
+    value: new FormControl(value),
+  }
+}
+
+type RatingFormControl = ReturnType<typeof createRatingFormControl>;
+
+export class MovieRatingForm extends FormEntity<RatingFormControl> {
+  constructor(rating?: Partial<MovieRating>) {
+    super(createRatingFormControl(rating));
+  }
+}
+
+export function createMovieRating(params: Partial<MovieRating> = {}): MovieRating {
+  return {
+    country: null,
+    value: '',
+    ...params
+  };
+}
+
+// ------------------------------
+//        ORIGINAL RELEASE
+// ------------------------------
+
+function createOriginalReleaseFormControl(entity?: Partial<MovieOriginalRelease>) {
+  const { country, date, media } = createMovieOriginalRelease(entity);
+  return {
+    country: new FormStaticValue(country, 'TERRITORIES'),
+    date: new FormControl(date),
+    media: new FormControl(media),
+  }
+}
+
+type OriginalReleaseFormControl = ReturnType<typeof createOriginalReleaseFormControl>;
+
+export class OriginalReleaseForm extends FormEntity<OriginalReleaseFormControl> {
+  constructor(originalRelease?: Partial<MovieOriginalRelease>) {
+    super(createOriginalReleaseFormControl(originalRelease));
+  }
+}
+
+export function createMovieOriginalRelease(
+  params: Partial<MovieOriginalRelease> = {}
+): MovieOriginalRelease {
+  return {
+    country: null,
+    ...params,
+    date: toDate(params.date),
+  };
+}
+
+// ------------------------------
+//          LANGUAGES
+// ------------------------------
+
+export function createLanguageControl(
+  versionInfo: Partial<{ [language in LanguagesSlug]: MovieLanguageSpecification }>
+) {
+  const controls = {};
+  for (const language in versionInfo) {
+    controls[language] = new VersionSpecificationForm(versionInfo[language]);
+  }
+  return controls;
+}
+
+export class MovieVersionInfoForm extends FormEntity<any> {
+  constructor(
+    versionInfo: Partial<{ [language in LanguagesSlug]: MovieLanguageSpecification }> = {}
+  ) {
+    super(createLanguageControl(versionInfo));
+  }
+
+  addLanguage(language: LanguagesSlug, value?: Partial<MovieLanguageSpecification>) {
+    const spec = createMovieLanguageSpecification(value);
+    this.setControl(language, new VersionSpecificationForm(spec));
+  }
+
+  removeLanguage(language: LanguagesSlug) {
+    this.removeControl(language);
+    this.updateValueAndValidity();
+  }
+}
+
+export class VersionSpecificationForm extends FormEntity<any> {
+  constructor(versionSpecifictaion: MovieLanguageSpecification) {
+    super({
+      original: new FormControl(versionSpecifictaion.original),
+      dubbed: new FormControl(versionSpecifictaion.dubbed),
+      subtitle: new FormControl(versionSpecifictaion.subtitle),
+      caption: new FormControl(versionSpecifictaion.caption)
     });
   }
 }
