--- libs/movie/src/lib/movie/form/review/review.form.ts
@@ -1,32 +1,0 @@
-import { FormControl } from '@angular/forms';
-import { MovieReview } from '../../+state/movie.firestore';
-import { FormEntity } from '@blockframes/utils/form/forms/entity.form';
-import { urlValidators } from '@blockframes/utils/form';
-
-function createMovieReviewControl(review: Partial<MovieReview> = {}) {
-  const { criticName, journalName, criticQuote, revueLink } = createMovieReview(review);
-  return {
-    criticName: new FormControl(criticName),
-    journalName: new FormControl(journalName),
-    criticQuote: new FormControl(criticQuote),
-    revueLink: new FormControl(revueLink, urlValidators),
-  }
-}
-
-export type MovieReviewControls = ReturnType<typeof createMovieReviewControl>;
-
-export class MovieReviewForm extends FormEntity<MovieReviewControls, MovieReview> {
-  constructor(review?: Partial<MovieReview>) {
-    super(createMovieReviewControl(review));
-  }
-}
-
-export function createMovieReview(params: Partial<MovieReview> = {}): MovieReview {
-  return {
-    criticName: '',
-    journalName: '',
-    criticQuote: '',
-    revueLink: '',
-    ...params,
-  }
-}
